+++ embedded_voting/epistemicGenerators/__init__.py
@@ -1,5 +1,5 @@
-from embedded_voting.epistemicGenerators.ratingsgeneratorepistemic import *
-from embedded_voting.epistemicGenerators.ratingsgeneratorepistemicgroupedmix import *
-from embedded_voting.epistemicGenerators.ratingsgeneratorepistemicgroupedmean import *
-from embedded_voting.epistemicGenerators.ratingsgeneratorepistemicgroupednoise import *
-from embedded_voting.epistemicGenerators.ratingsgeneratorepistemicmultivariate import *
+from embedded_voting.epistemicGenerators.ratings_generator_epistemic import *
+from embedded_voting.epistemicGenerators.ratings_generator_epistemic_grouped_mix import *
+from embedded_voting.epistemicGenerators.ratings_generator_epistemic_grouped_mean import *
+from embedded_voting.epistemicGenerators.ratings_generator_epistemic_grouped_noise import *
+from embedded_voting.epistemicGenerators.ratings_generator_epistemic_multivariate import *
