+++ embedded_voting/epistemicGenerators/ratings_generator_epistemic_grouped_noise.py
@@ -1,14 +1,12 @@
 import numpy as np
-from embedded_voting.epistemicGenerators.ratingsgeneratorepistemic import RatingsGeneratorEpistemic
+from embedded_voting.epistemicGenerators.ratings_generator_epistemic import RatingsGeneratorEpistemic
 from embedded_voting.ratings.ratings import Ratings
 
 
-class RatingsGeneratorEpistemicGroupedMean(RatingsGeneratorEpistemic):
+class RatingsGeneratorEpistemicGroupedNoise(RatingsGeneratorEpistemic):
     """
-    A generator of scores such that voters are
-    separated into different groups and the noise of
-    an voter on an alternative is equal to the noise
-    of his group plus his own independent noise.
+    A generator of scores such that voters are separated into different groups and for each
+    alternative the variance of each voter of the same group is the same.
 
     Parameters
     ----------
@@ -16,9 +14,7 @@ class RatingsGeneratorEpistemicGroupedMean(RatingsGeneratorEpistemic):
         The number of voters in each groups.
         The sum is equal to :attr:`~embedded_voting.ScoreGenerator.n_voters`.
     group_noise : float
-        The variance used to sample the noise of each group.
-    independent_noise : float
-        The variance used to sample the independent noise of each voter.
+        The variance used to sample the variances of each group.
     minimum_score : float or int
         The minimum true value of an alternative.
         By default, it is set to 10.
@@ -34,21 +30,20 @@ class RatingsGeneratorEpistemicGroupedMean(RatingsGeneratorEpistemic):
 
     Examples
     --------
-    >>> np.random.seed(44)
-    >>> generator = RatingsGeneratorEpistemicGroupedMean([2, 2])
+    >>> np.random.seed(42)
+    >>> generator = RatingsGeneratorEpistemicGroupedNoise([2, 2])
     >>> generator()
-    Ratings([[19.07101495],
-             [19.07101496],
-             [19.27467426],
-             [19.27467425]])
+    Ratings([[14.03963831],
+             [14.81094637],
+             [13.41737103],
+             [13.44883031]])
     >>> generator.ground_truth_
-    array([18.34842149])
+    array([13.74540119])
     """
-    def __init__(self, groups_sizes, group_noise=1, independent_noise=0, minimum_score=10, maximum_score=20):
+    def __init__(self, groups_sizes, group_noise=1, minimum_score=10, maximum_score=20):
         super().__init__(minimum_score=minimum_score, maximum_score=maximum_score,
                          groups_sizes=groups_sizes)
         self.group_noise = group_noise
-        self.independent_noise = independent_noise
 
     def __call__(self, n_candidates=1, *args):
         self.ground_truth_ = self.generate_true_values(n_candidates=n_candidates)
@@ -58,9 +53,8 @@ class RatingsGeneratorEpistemicGroupedMean(RatingsGeneratorEpistemic):
             cov = np.zeros((self.n_voters, self.n_voters))
             s = 0
             for k in range(len(self.groups_sizes)):
-                cov[s:s + self.groups_sizes[k], s:s + self.groups_sizes[k]] = np.ones(self.groups_sizes[k]) * sigma[k]
+                cov[s:s + self.groups_sizes[k], s:s + self.groups_sizes[k]] = np.eye(self.groups_sizes[k]) * sigma[k]
                 s += self.groups_sizes[k]
-            scores_i = np.random.multivariate_normal(np.ones(self.n_voters) * self.ground_truth_[i], cov)
-            scores_i += np.random.randn(self.n_voters) * self.independent_noise
+            scores_i = np.random.multivariate_normal(np.ones(self.n_voters)*self.ground_truth_[i], cov)
             scores[:, i] = scores_i
         return Ratings(scores)
