+++ embedded_voting/epistemicGenerators/ratings_generator_epistemic_linear_gaussian.py
@@ -1,5 +1,5 @@
 import numpy as np
-from embedded_voting.epistemicGenerators.ratingsgeneratorepistemic import RatingsGeneratorEpistemic
+from embedded_voting.epistemicGenerators.ratings_generator_epistemic import RatingsGeneratorEpistemic
 from embedded_voting.ratings.ratings import Ratings
